daos hechos

+++ gripeA_2020/dao/daoBrotes.py
@@ -1,5 +1,5 @@
-import pymongo
 from dao.daoMongo import daoMongo
+from transfer.transferBrote import transferBrote
 
 class daoBrotes(daoMongo):
 
@@ -7,20 +7,35 @@ class daoBrotes(daoMongo):
         super()
         self._doc = self._collection.outbreaks
 
-    def find(self, query = {}):
-        return self._doc.find(query)
+    def read(self, ident = []):
+        ret = []
+        query = {}
+        if len(ident) > 0:
+            query['oieid'] = {'$in' : ident}
 
-    def insert_one(self, transfer = {}):
-        return self._doc.insert_one(transfer)
-    def insert_many(self, transfer = {}):
-        return self._doc.insert_many(transfer)
+        cursor = self._doc.find(query)
+        for it in cursor:
+            t = transferBrote(it)
+            ret.append(t)
 
-    def update_one(self, query = {}, newValue = {}):
-        return self._doc.update_one(query, newValue)
-    def update_many(self, query = {}, newValue = {}):
-        return self._doc.update_many(query, newValue)
+        return ret
 
-    def delete_one(self, query = {}):
-        return self._doc.delete_one(query)
-    def delete_many(self, query = {}):
-        return self._doc.delete_many(query)
+    def create(self):
+        raise Exception("Create de daoBrotes")
+
+    def update(self):
+        raise Exception("Update de daoBrotes")
+
+    def delete(self, ident = []):
+        ret = []
+        query = {}
+        if len(ident) > 0:
+            query['oieid'] = {'$in' : ident}
+            delet = self._doc.delete_many(query)
+            return delet.deleted_count()
+        else:
+            return 0
+
+    def delete_all(self):
+        delet = self._doc.delete_many({})
+        return delet.deleted_count()
